+++ src/main/java/kitchenpos/domain/Product.java
@@ -1,5 +1,7 @@
 package kitchenpos.domain;
 
+import kitchenpos.domain.vo.Money;
+
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
