+++ src/main/java/kitchenpos/domain/vo/NumberOfGuest.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.domain.vo;
 
 import java.util.Objects;
